requirements.txt
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.6
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-guardian==1.4.9
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
PyJWT==1.5.3
PyMySQL==0.7.11
pytz==2017.3
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
